travel_ai_api/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
amadeus==9.0.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
Django==5.2
djangorestframework==3.16.0
duckduckgo_search==8.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
lxml==5.4.0
openai==1.77.0
primp==0.15.0
psycopg2-binary==2.9.9
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
